refactor: change files structure, added Header component, added props in card and header
