+++ t2.01/src/App.css
@@ -1,0 +1,7 @@
+.container {
+  max-width: 900px;
+  margin: 0 auto;
+  padding: 0 15px;
+  border: 1px solid black;
+  min-height: 500px;
+}
