+++ t2.01/src/App.js
@@ -1,21 +1,13 @@
 import React from "react";
-import './css/main.css'
-import Card from "./components/Card";
+import './App.css'
+import Card from "./components/Card/Card";
+import Header from "./components/Header/Header";
 
 function App() {
   return (
     <div className="container">
-      <header className="header">
-        <h1 className="sr-only">Cards</h1>
-        <h2 className="header__title">Header</h2>
-      </header>
-      <main className="main">
-        <section className="content">
-          <ul className="content__items">
-            <Card />
-          </ul>
-        </section>
-      </main>
+      <Header title="Header"/> 
+      <Card title="card" text="card text"/>
     </div>
   );
 }
