--- t2.01/src/components/Card.js
@@ -1,13 +1,0 @@
-import React from "react";
-import './Card.css'
-
-const Card = () => {
-  return (
-    <li className="content__item">
-      <h2 className="content__title">Card</h2>
-      <p className="content__text">Card text</p>
-    </li>
-  );
-}
-
-export default Card;
