+++ t2.01/src/components/Header/Header.css
@@ -1,0 +1,10 @@
+.header {
+  padding: 10px 30px 40px 30px;
+  text-align: center;
+}
+
+.header__title {
+  font-size: 50px;
+  letter-spacing: 1px;
+  color: red;
+}
