--- t2.01/src/css/basic.css
@@ -1,18 +1,0 @@
-.container {
-  max-width: 900px;
-  margin: 0 auto;
-  padding: 0 15px;
-  border: 1px solid black;
-  min-height: 500px;
-}
-
-.sr-only {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0,0,0,0);
-  border: 0;
-}
