--- t2.01/src/css/main.css
@@ -1,20 +1,0 @@
-.header {
-  padding: 10px 30px;
-  text-align: center;
-}
-
-.header__title {
-  font-size: 50px;
-  letter-spacing: 1px;
-  color: red;
-}
-
-.content {
-  padding: 20px 30px;
-}
-
-.content__items {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 30px;
-}
